travel_ai/requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv==1.1.1
httpx
numpy
scikit-learn